+++ demo/src/js/demo.js
@@ -179,7 +179,7 @@ import Raven from 'raven-js';
                         player.source = {
                             type: 'video',
                             sources: [{
-                                src: 'https://vimeo.com/76979871',
+                                src: 'https://vimeo.com/25345658',
                                 provider: 'vimeo',
                             }],
                         };